Refactored code - Extracted floating logic to separate hook - Updated styles
