+++ src/components/edit_area/hover_ghost.less
@@ -1,0 +1,6 @@
+.timeline-editor {
+  &-hover-ghost {
+    height: 100%;
+    pointer-events: none;
+  }
+}
